initial - new splash page & reorganized nav
