+++ src/pages/wa/index.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import ReactPlayer from "react-player";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+import Countdown from "react-countdown";
 
 import {
   NavBar,
@@ -27,66 +28,97 @@ import {
   PageContainer,
   VidSection,
   NavSection,
-  SorryBanner,
-  KoloheBanner,
+  HeaderImg,
+  TimeWOz,
+  HostShot,
+  CountCountainer,
+  EventBanner,
   BuyContainer,
 } from "./styles";
 
-import Modal from "../../components/modal/Modal";
-import useModal from "../../components/modal/useModal";
+import kael from "../../lib/assets/kael.jpg";
+import jack from "../../lib/assets/jack.jpg";
+import jay from "../../lib/assets/jay.jpg";
+import jacob from "../../lib/assets/jacob.jpg";
 
-import kolohe from "../../lib/assets/Kolohe.png";
-import griffin from "../../lib/assets/Griffin.png";
-import ian from "../../lib/assets/Ian.png";
+import yadin from "../../lib/assets/yadin.jpg";
 import dane from "../../lib/assets/Dane.png";
-import selema from "../../lib/assets/Selema.png";
+import selema from "../../lib/assets/selema.jpg";
 import taylor from "../../lib/assets/Taylor.png";
 import s100white from "../../lib/assets/s100logowhite.png";
 import s100dark from "../../lib/assets/s100logodark.png";
 import insta from "../../lib/assets/insta.png";
+import imgheader from "../../lib/assets/site_header.jpg";
 
 const SplashPage = () => {
-  const { isShowing, toggle } = useModal();
   const [email, setEmail] = useState("");
   const [loading, setLoading] = useState(true);
   const [isRegistered, setIsRegistered] = useState(false);
 
+  const Completionist = () => (
+    <div className="countdown">
+      {/* <span>Show time</span> : <span>now</span> */}
+    </div>
+  );
+
+  const renderer = ({ days, hours, minutes, seconds, completed }) => {
+    if (completed) {
+      // Render a completed state
+      return <Completionist />;
+    } else {
+      // Render a countdown
+      return (
+        <div class="countdown">
+          <p>
+            <span>Show time</span> : <span>{days} Days</span> :{" "}
+            <span>{hours} Hours</span> : <span>{minutes} Minutes</span> :{" "}
+            <span>{seconds} Seconds</span>
+          </p>
+        </div>
+      );
+    }
+  };
+
   return (
     <PageContainer>
       <NavSection>
-        {" "}
-        <Banner></Banner>
         <NavBar>
           <a href="/">
             <img src={s100dark} />
           </a>
-          {/* <div>
-            <Button href="/event">WA</Button>
-          </div> */}
+          <CountCountainer>
+            <Countdown date={"09/17/2020 5:00 PM PST"} renderer={renderer} />
+          </CountCountainer>{" "}
           <BuyContainer>
             <Button href="/wa">WA</Button>
             <Button href="/trestles">TRESTLES</Button>
             <Button href="/event">Event</Button>
           </BuyContainer>
         </NavBar>
+        <Banner>
+          <Countdown date={"09/17/2020 5:00 PM PST"} renderer={renderer} />
+        </Banner>
       </NavSection>
-      <KoloheBanner>
-        <a href="https://stabmag.com/stabcinema/surf100-lower-trestles/">
-          View Surf100 Lowers 
+
+      <EventBanner>
+        <a href="https://stabmag.com/stabcinema/surf100-western-australia/">
+          View Surf100 West Oz
         </a>
-      </KoloheBanner>
+      </EventBanner>
+
+      <HeaderImg src={imgheader}></HeaderImg>
 
       <ContentContainer>
         <TitleSection>
           <h2>An audience-judged, 100-minute wave riding demonstration.</h2>
-          <h2>6pm (PST) August 6, 2020</h2>
+          <h2>6pm (PST) September 17, 2020</h2>
         </TitleSection>
         <VidSection>
           <VidContainer>
             {" "}
             <ReactPlayer
               className="plyr"
-              url="https://youtu.be/F-uIHXwf87I"
+              url="https://www.youtube.com/watch?v=CvaBHhiKddY&feature=youtu.be"
               width="100%"
             />
           </VidContainer>
@@ -97,38 +129,44 @@ const SplashPage = () => {
           <h2>Featuring:</h2>
           <HeadShotContainer>
             <ProfileContainer>
-              <HeadShot src={kolohe}></HeadShot>
-              <h5>Kolohe Andino</h5>
+              <HeadShot src={kael}></HeadShot>
+              <h5>Kael Walsh</h5>
+              <p>(in yellow)</p>
+            </ProfileContainer>
+            <ProfileContainer>
+              <HeadShot src={jack}></HeadShot>
+              <h5>Jack Robinson</h5>
+              <p>(in red)</p>
+            </ProfileContainer>
+            <ProfileContainer>
+              <HeadShot src={jay}></HeadShot>
+              <h5>Jay Davies</h5>
               <p>(in black)</p>
             </ProfileContainer>
             <ProfileContainer>
-              <HeadShot src={griffin}></HeadShot>
-              <h5>Griffin Colapinto</h5>
-              <p>(in orange)</p>
-            </ProfileContainer>
-            <ProfileContainer>
-              <HeadShot src={ian}></HeadShot>
-              <h5>Ian Crane</h5>
+              <HeadShot src={jacob}></HeadShot>
+              <h5>Jacob Willcox</h5>
               <p>(in blue)</p>
             </ProfileContainer>
           </HeadShotContainer>
         </ContentContainer>
       </Featuring>
+
       <HostedSection>
         <ContentContainer>
           <h2>Hosted by:</h2>
-          <HeadShotContainer>
+          <HeadShotContainer className="hosted">
             <ProfileContainer>
-              <HeadShot src={dane}></HeadShot>
+              <HostShot src={dane}></HostShot>
               <h5>Dane Reynolds</h5>
             </ProfileContainer>
             <ProfileContainer>
-              <HeadShot src={selema}></HeadShot>
+              <HostShot src={selema}></HostShot>
               <h5>Selema Masekela</h5>
             </ProfileContainer>
             <ProfileContainer>
-              <HeadShot src={taylor}></HeadShot>
-              <h5>Taylor Knox</h5>
+              <HostShot src={yadin}></HostShot>
+              <h5>Yadin Nicol</h5>
             </ProfileContainer>
           </HeadShotContainer>
         </ContentContainer>
@@ -136,36 +174,27 @@ const SplashPage = () => {
 
       <ContentContainer>
         <Description>
-          <h2>What is Surf 100?</h2>
+          <h2>What is Surf100?</h2>
           <ul>
             <p>
-              Surf 100 is a waveriding exposition that takes place over a
-              single, 100-minute period and is judged live, at a later date, by
-              the internet. All of this is done following social distancing
-              guidelines. San Clemente favorite sons Kolohe Andino, Griffin
-              Colapinto, and Ian Crane will paddle out for a 100-minute
-              waveriding demonstration at Lower Trestles.
+              Surf100 is a waveriding exposition that takes place over a single,
+              100-minute period and is judged live, at a later date, by the
+              internet. All of this is done following social distancing
+              guidelines. In this edition, West Oz mainstays Jack Robinson, Jay
+              Davies, Jacob Willcox, and Kael Walsh will paddle out for a
+              100-minute waveriding demonstration at North Point.
             </p>
             <p>
-              The surfers can ride as many waves as they want in the 100-minute
-              session. Each wave will be considered a separate “clip” that can
-              be used in their final “section”. Each section will consist of two
-              clips—one right, one left—which will be judged out of 100 points
-              apiece, 200 points total. (See above for more info on judging.)
-            </p>
-            <p>
-              While negotiating a typical Lowers crowd, they’ll attempt to
-              capture the best two-wave (right + left) video clip in the
-              allotted time.
-            </p>
-            <p>
-              The best section (two clip total), according to the internet,
+              While negotiating a typical North Point crowd, the surfers can
+              ride as many waves as they want in the 100-minute session. Each
+              wave will be scored by the audience out of 100 points apiece. The
+              surfer with the best two-wave total, according to the internet,
               wins.
             </p>
             <p>
-              The show will be anchored with the no holds-barred team of
-              analysts and thought leaders, Dane Reynolds, Taylor Knox and
-              Selema Masekela.
+              The show will be anchored by the no holds-barred team of analysts
+              and thought leaders, Dane Reynolds, Selema Masekela, and Yadin
+              Nicol.
             </p>
           </ul>
         </Description>
@@ -174,11 +203,14 @@ const SplashPage = () => {
         <ShowContent>
           <h2>Show Time:</h2>
           <TimeUs>
-            <p>6pm, Thursday August 6, California (PST)</p>
+            <p>6pm, Thursday September 17, California (PST)</p>
           </TimeUs>
           <TimeOz>
-            <p>11am, Friday August 7, QLD/NSW/VIC (AEST)</p>
+            <p>11am, Friday September 18, QLD/NSW/VIC (AEST)</p>
           </TimeOz>
+          <TimeWOz>
+            <p>9 am, Friday September 18, West Oz (AWST)</p>
+          </TimeWOz>
         </ShowContent>
       </ShowTime>
       <FaqSection id="faq">
@@ -187,29 +219,15 @@ const SplashPage = () => {
           <ColumnOne>
             <h4>When is the event live?</h4>
             <p>
-              6pm, Thursday August 6, California (PST) & 11am, Friday August 7,
-              QLD/NSW/VIC (AEST){" "}
+              6pm, Thursday September 17, California (PST) & 11am, Friday
+              September 18, QLD/NSW/VIC (AEST)
             </p>
-
-            <h4>Are there going to be more events?</h4>
-            <p>
-              Pending the success of the first event, we plan on shooting this
-              in different regions, globally.
-            </p>
-
-            <h4>Does it cost money to download the app?</h4>
-            <p>There’s no app, all viewing takes place on the web browser.</p>
 
             <h4>How long is the broadcast?</h4>
             <p>
-              110 minutes. The 100 minute event plus some pre and post show.
+              110 minutes. The 100 minute event (played in “real-time”) plus
+              some pre and post show.
             </p>
-
-            <h4>
-              I’m on a phone so only have the Safari browser, will I be able to
-              watch Surf 100?
-            </h4>
-            <p>Yes.</p>
 
             <h4>Why are you showing the event after the fact? And not live?</h4>
             <p>
@@ -223,7 +241,7 @@ const SplashPage = () => {
             </h4>
             <p>
               You can watch and score at the same time. Obvs, you can’t view
-              full screen.
+              full screen
             </p>
 
             <h4>Can I watch on one device and score on another?</h4>
@@ -234,11 +252,8 @@ const SplashPage = () => {
 
             <h4>How do I pre-order the pay per view?</h4>
             <p>
-              <a href="/trestles">Click here.</a>
+              <a href="/event">Click here.</a>
             </p>
-
-            <h4>Does it cost more to get the pay per view in HD.</h4>
-            <p>No.</p>
 
             <h4>
               Oh, and does this $USD14,99 get me the season or just one show?
@@ -266,17 +281,9 @@ const SplashPage = () => {
             <h4>What does the winner win?</h4>
             <p>Bragging rights.</p>
 
-            <h4>Is there a separate contest for the viewers who judge?</h4>
-            <p>
-              We have our head critic, Mikey Ciaramella, reveal his scores after
-              our audience scores are locked. At the end of the event, the judge
-              with the same or nearest score to Mikey will win a three-board
-              Mayhem quiver.{" "}
-            </p>
-
             <h4>Can we judge as a group?</h4>
             <p>
-              No, singular judging only sorry. Everyone can judge individually.
+              No, singular judging only, sorry. Everyone can judge individually.
             </p>
 
             <h4>What do I do if my video doesn’t work?</h4>
@@ -285,13 +292,13 @@ const SplashPage = () => {
               athletic surfer from Virginia, will help you through it.
             </p>
 
+            <h4>Does Dane Reynolds judge?</h4>
+            <p>No. He’s on commentary.</p>
+
             <h4>
               What if I’m late to the screening? Do I start from the start?
             </h4>
             <p>You’ll start mid-broadcast.</p>
-
-            <h4>Can I watch on my phone and judge at the same time?</h4>
-            <p>Yup. Your log-in will work on two devices.</p>
 
             <h4>Can I watch it on my smart TV?</h4>
             <p>
@@ -304,10 +311,8 @@ const SplashPage = () => {
 
             <h4>Do I have to judge every wave?</h4>
             <p>
-              You don’t but our aim is to have consistent judging so we hope you
-              can so we can have as large a sample size as possible. Also,
-              missing a wave makes it significantly less likely you’ll win the
-              ...Lost quiver.
+              Not the “irrelevant waves” with insignificant scores. But every
+              wave of note you will have to score.
             </p>
 
             <h4>
@@ -327,6 +332,7 @@ const SplashPage = () => {
           </ColumnTwo>
         </FaqContainer>
       </FaqSection>
+
       <Footer>
         <a href="/">
           <img src={s100white} />
